refactor: refact distributed transaction filter support comQuery
